+++ idempotency_constants.go
@@ -1,0 +1,6 @@
+package chargebee
+
+const (
+	IdempotencyHeader       = "chargebee-idempotency-key"
+	IdempotencyReplayHeader = "Chargebee-Idempotency-Replayed"
+)